Fix bug in unfreeze
The unfreeze_object function takes a dict as its argument and
it was modifying the contents of the dict (as a mutable arg).
It now copies the dict before processing it so that the original
is maintained. This fixes a bug with deep serialization where
the meta node was being stripped out and so the deserialization
process failed thinking that the field was a dict, not a frozen
object.